pyproject.toml
[tool.poetry]
name = "demotivator-bot"
version = "0.1.1"
description = "Супер-пупер-мега демотиватор-бот для гифок на стероидах и FFmpeg."
authors = ["Daniel Kotofiev <nesclass@yandex.ru>"]

[tool.poetry.dependencies]
python = "^3.11"
aiogram = "^2.24"
ffmpeg-python = "^0.2.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
